backend/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
